Balance sheet added cash on hand

+++ Spring_Boot_Balaji_FInance/sri-balaji-finance/src/main/java/com/balaji/finance/dto/BalanceSheetProjection.java
@@ -1,9 +1,11 @@
 package com.balaji.finance.dto;
 
+import java.math.BigDecimal;
+
 public interface BalanceSheetProjection {
 
     String getType();
     String getMasterCode();
     String getCode();
-    Double getAmount();
+    BigDecimal getAmount();
 }
