+++ Spring_Boot_Balaji_FInance/sri-balaji-finance/src/main/java/com/balaji/finance/dto/RevenueExpenseProjection.java
@@ -1,8 +1,10 @@
 package com.balaji.finance.dto;
 
+import java.math.BigDecimal;
+
 public interface RevenueExpenseProjection {
 
     String getType();
     String getCode();
-    Double getAmount();
+    BigDecimal getAmount();
 }
